WorkItem: 4012 Implemented unshare dialog + action
The changes generally take a few seconds to take into effect when turning
sharing on/off. Do it too fast and Agilefant will still think you're in
one state or another. Not sure if that's considered a bug or not.

JUint tests all cleared and whatnot.

Tomorrow (Thursday March 15 2012) I'm going to review what I've added at
the next standup to make sure it's to standard. Please check for bugs!